Add support for structured logging
Add a config option which supports a value of `json` in order to enable
structured logging using the JSON format. This can be done by setting
the `logger.format` option in the environment specific configuration
file, or by specifying the `logger__format` environment variable.

An example log line will look like:

```json
{"message":"[development] Hearth FHIR server running on 0.0.0.0:3447","level":"info","timestamp":"2021-08-23T08:03:46.293Z"}
```

This allows for easier filtering of logs when using tools which support
line-delimited JSON such as `jq` or Amazon CloudWatch.